# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  notification.CreateSMTPConfigRequest:
    properties:
      from_email:
        type: string
      from_name:
        type: string
      host:
        type: string
      organization_id:
        type: string
      password:
        type: string
      port:
        type: integer
      username:
        type: string
    type: object
  notification.DeleteSMTPConfigRequest:
    properties:
      id:
        type: string
    type: object
  notification.GetPreferencesResponse:
    properties:
      activity:
        items:
          $ref: '#/definitions/notification.PreferenceType'
        type: array
      security:
        items:
          $ref: '#/definitions/notification.PreferenceType'
        type: array
      update:
        items:
          $ref: '#/definitions/notification.PreferenceType'
        type: array
    type: object
  notification.PreferenceType:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      label:
        type: string
    required:
    - description
    - id
    - label
    type: object
  notification.UpdatePreferenceRequest:
    properties:
      category:
        enum:
        - activity
        - security
        - update
        type: string
      enabled:
        type: boolean
      type:
        type: string
    required:
    - category
    - type
    type: object
  notification.UpdateSMTPConfigRequest:
    properties:
      from_email:
        type: string
      from_name:
        type: string
      host:
        type: string
      id:
        type: string
      password:
        type: string
      port:
        type: integer
      username:
        type: string
    type: object
...
  types.CreateDomainRequest:
    properties:
      name:
        type: string
      organization_id:
        type: string
    type: object
  types.CreateDomainResponse:
    properties:
      id:
        type: string
    type: object
...
  types.DeleteDomainRequest:
    properties:
      id:
        type: string
    type: object
...
  types.Domain:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      organization_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  types.RandomSubdomainResponse:
    properties:
      domain:
        type: string
      subdomain:
        type: string
    type: object
...
  types.Response:
    description: Standard API response structure
    properties:
      data:
        description: Actual response data
      error:
        description: |-
          Error message in case of error
          @Example Invalid input parameters
        type: string
      message:
        description: |-
          Optional message providing additional information
          @Example Operation completed successfully
        type: string
      status:
        description: |-
          Status of the response ("success" or "error")
          @Example success
        type: string
    type: object
...
  types.UpdateDomainRequest:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: raghav@nixopus.com
    name: API Support
    url: http://www.swagger.io/support
  description: Api for Nixopus
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://nixopus.com/terms/
  title: Nixopus Documentation
  version: "1.0"
paths:
...
  /domain:
    delete:
      consumes:
      - application/json
      description: Deletes a domain by its ID.
      parameters:
      - description: Domain deletion request
        in: body
        name: deleteDomain
        required: true
        schema:
          $ref: '#/definitions/types.DeleteDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Domain deleted successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Delete a domain
      tags:
      - domain
    post:
      consumes:
      - application/json
      description: Creates a new domain in the application.
      parameters:
      - description: Domain creation request
        in: body
        name: domain
        required: true
        schema:
          $ref: '#/definitions/types.CreateDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with domain
          schema:
            $ref: '#/definitions/types.CreateDomainResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Create a new domain
      tags:
      - domain
    put:
      consumes:
      - application/json
      description: Updates an existing domain in the application.
      parameters:
      - description: Domain update request
        in: body
        name: domain
        required: true
        schema:
          $ref: '#/definitions/types.UpdateDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with updated domain
          schema:
            $ref: '#/definitions/types.Domain'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Update a domain
      tags:
      - domain
  /domain/generate:
    get:
      consumes:
      - application/json
      description: Generates a random subdomain by taking a random domain from the
        list of all domains and appending a random string to it.
      produces:
      - application/json
      responses:
        "200":
          description: Success response with random subdomain
          schema:
            $ref: '#/definitions/types.RandomSubdomainResponse'
        "404":
          description: No domains available
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Generate a random subdomain
      tags:
      - domain
  /domains:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all domains.
      produces:
      - application/json
      responses:
        "200":
          description: Success response with domains
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Get all domains
      tags:
      - domain
...
  /notification/preferences:
    get:
      consumes:
      - application/json
      description: Retrieves the notification preferences for the current user
      produces:
      - application/json
      responses:
        "200":
          description: Success response with preferences
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/notification.GetPreferencesResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Get notification preferences
      tags:
      - notification
    post:
      consumes:
      - application/json
      description: Update notification preference
      parameters:
      - description: Notification preference
        in: body
        name: preference
        required: true
        schema:
          $ref: '#/definitions/notification.UpdatePreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Notification preferences updated successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Update notification preference
      tags:
      - notification
  /notification/smtp:
    delete:
      consumes:
      - application/json
      description: Delete SMTP configuration
      parameters:
      - description: SMTP configuration
        in: body
        name: SMTPConfigs
        required: true
        schema:
          $ref: '#/definitions/notification.DeleteSMTPConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: SMTP deleted successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Delete SMTP configuration
      tags:
      - notification
    get:
      consumes:
      - application/json
      description: Get SMTP configuration
      produces:
      - application/json
      responses:
        "200":
          description: SMTP configuration
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Get SMTP configuration
      tags:
      - notification
    post:
      consumes:
      - application/json
      description: Add SMTP configuration
      parameters:
      - description: SMTP configuration
        in: body
        name: SMTPConfigs
        required: true
        schema:
          $ref: '#/definitions/notification.CreateSMTPConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: SMTP added successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Add SMTP configuration
      tags:
      - notification
    put:
      consumes:
      - application/json
      description: Update SMTP configuration
      parameters:
      - description: SMTP configuration
        in: body
        name: SMTPConfigs
        required: true
        schema:
          $ref: '#/definitions/notification.UpdateSMTPConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: SMTP updated successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Update SMTP configuration
      tags:
      - notification
...
securityDefinitions:
  BearerAuth:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"